+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Unreleased
+----------
+- Spawn `chromedriver` and derived processes into dedicated process
+  group
+
+
 0.1.1
 -----
 - Made sure to collect zombie processes after termination